Exploring the Natural Park
This project includes five main teleport stations in which there are different interactions for players to do. This VR experience designed for Desktop VR (without headset) and works with Arrow Keys and Gaze-Mouse, giving people an opportunity to explore a natural fantasy environment.
This prototype can be developed for an IVR experiment to examine how it will improve children’s soft skills, designed various activities regarding these skills. A player as a child can learn, practice, and get instant feedback about their performance. This experience can be integrated with an AI assistant who mentors children during the process, giving a piece of advice on how to interact with people and how to be a better team player.